feat: Prepare scripts for npm publish
