+++ index.js
@@ -1,0 +1,8 @@
+require('dotenv').config()
+const ExcelFile = require('./src/classes/excel')
+const ExcelFactory = require('./src/classes/excelfactory')
+
+module.exports = {
+  ExcelFile,
+  ExcelFactory
+}
